requirements.txt
# API and HTTP
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0

# LLM and Embeddings
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
sentence-transformers==2.2.2

# Vector Database
qdrant-client==1.7.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2 